smart-resume-backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
python-docx
spacy
transformers
torch
pandas
python-dotenv
requests
PyPDF2
